feat(GuestRepositoryImpl.java): add EventJpaRepository to save event with guest refactor(GuestEntity.java): remove id from equals and hashCode methods refactor(GuestServiceImpl.java): move to ports.input package feat(application.yml): add aop.proxy-target-class configuration
The EventJpaRepository was added to the GuestRepositoryImpl to allow saving an event along with a guest. This change was necessary to ensure that each guest is associated with an event.

The equals and hashCode methods in GuestEntity were refactored to remove the id field. This was done to ensure that two GuestEntity objects are considered equal if their non-id fields are equal, which is more in line with the business logic.

The GuestServiceImpl was moved to the ports.input package to better organize the code and make it easier to understand the architecture.

The aop.proxy-target-class configuration was added to the application.yml file to control whether CGLIB proxies should be created. This was done to improve the performance of the application by avoiding the creation of unnecessary proxies.